Create Reverse Linked List in groups of Size K (#156)
I have created a solution file to the Problem Statement: Given the head of a singly linked list of n nodes and an integer k, where k is less than or equal to n. Your task is to reverse the order of each group of k consecutive nodes, if n is not divisible by k, then the last group of remaining nodes should remain unchanged.